+++ xarray/backends/api.py
@@ -490,8 +490,7 @@ def open_dataset(
           other attributes as coordinate variables.
 
         Only existing variables can be set as coordinates. Missing variables
-        will be silently ignored. If ``coordinates`` do not match any variable,
-        no decoding takes place.
+        will be silently ignored. 
     drop_variables: str or iterable of str, optional
         A variable or list of variables to exclude from being parsed from the
         dataset. This may be useful to drop variables with problems or
@@ -697,8 +696,7 @@ def open_dataarray(
           other attributes as coordinate variables.
 
         Only existing variables can be set as coordinates. Missing variables
-        will be silently ignored. If ``coordinates`` do not match any variable,
-        no decoding takes place.
+        will be silently ignored.
     drop_variables: str or iterable of str, optional
         A variable or list of variables to exclude from being parsed from the
         dataset. This may be useful to drop variables with problems or